add missing playerId type definition to protocol21029.py
Discovered when I updated my (ancient) library to the latest and could no
longer decode the game events in one of my saved replays. It seems that at
some point last year these protocol files were regenerated, and the resulting
file for 21029 (and possibly others) have set the replay_userid_typeid to
None, and the corresponding struct definition is missing. This is a big
problem, because the userIds can still be found in the "replay.game.events"
stream, but now the decoders have no way to decode or skip over them (and
then incorrectly decode the remaining bits). Also worth noting is that the
type *can* still be found in the protocol21029.json schema definition.

Fixed by adding the type back in, by comparing with my old version of the
protocol. It has one field dependent on a 5-bit "_int" type, which can be
found as entry 37 in the type list. It's entirely possible other protocol
versions are affected, and if these files are automatically generated, then
this is probably not the preferred approach to fixing this type of omission.