PR #82: Add missing playerId type definition to protocol21029.py (#82)
Accepting this fix. When I was last working on this I re-did the code generator pass to pull all legacy protocol definitions and apparently it still missed this. Unfortunately, I don't have a way of verifying this fix or fixes against many other legacy protocols outside of people such as @paulfred submitting fixes. So thank you!